Fix compilation issues (#431)
* feat: remove deprecated files

* fix: library ignored unix compile options, issue #422

* fix: update deprecated syscalls

* fix: previously ignored compilation errors

* fix: bad function return type

* fix: cast strtoul return value

* fix: avoid mask instruction

* fix: ptrdiff_t casting

* fix: unsigned constants

* fix: ext size calculation

* fix: ext number calculation

* fix: internal int type casting

* fix: unix system type casting

* chore: run clang

* fix: windows compilation

* fix: usleep arg type cast

* fix: set lease factor as int

* fix: lease substraction on unsigned type

* build: use debug config by default with cmake

* ci: switch tests to dev/1.0.0 zenoh branch

* fix: missing end of file

* fix: replace ssize_t

* fix: missing end of file newline

* build: don't set -Wpedantic on unix c99

* fix: missing end of file newline

* fix: bad function prototypes

* fix: bad return function

* fix: bad function prototype

* fix: missing end of file newline

* fix: remove linux specific type cast

* fix: raweth & single thread compilation issues

* fix: modularity compilation issues

* fix: raweth compilation issues

* fix: api test compiler issue

* fix: sub memory leak

* fix: remove rebase shenanigans

* fix: empty struct error

+++ include/zenoh-pico/system/link/raweth.h
@@ -87,8 +87,8 @@ size_t _z_send_raweth(const _z_sys_net_socket_t *sock, const void *buff, size_t 
 size_t _z_receive_raweth(const _z_sys_net_socket_t *sock, void *buff, size_t buff_len, _z_slice_t *addr,
                          const _zp_raweth_whitelist_array_t *whitelist);
 int8_t _z_close_raweth(_z_sys_net_socket_t *sock);
-size_t _z_raweth_ntohs(size_t val);
-size_t _z_raweth_htons(size_t val);
+uint16_t _z_raweth_ntohs(uint16_t val);
+uint16_t _z_raweth_htons(uint16_t val);
 
 #endif
